Add mermaid diagram support
Use it in a code block like this

    ```mermaid
    flowchart TD
      Start --> Stop
    ```

Add mermaid, webpack, webpack-cli as a node devDependency.

Add a node script to build the mermaid bundle in a 'dist' subdirectory.
That subdirectory is ignored by git.

Add a mermaid 'partial' to copy those bundle files out to the
published static site.  And then the live JavaScript snippet
pulls in the mermaid module to draw the graph.